refactor: Change implementation navigation header.
