+++ src/admin/components/layouts/Main/index.tsx
@@ -1,5 +1,5 @@
-import Header from '@admin/components/elements/Header/Main';
 import Nav from '@admin/components/elements/Nav';
+import NavHeader from '@admin/components/elements/NavHeader';
 import { Box, Toolbar } from '@mui/material';
 import React, { useState } from 'react';
 import { Outlet } from 'react-router-dom';
@@ -14,10 +14,10 @@ const MainLayout: React.FC<Props> = ({ group }) => {
 
   return (
     <Box sx={{ display: 'flex', width: '100%' }}>
-      <Header toggleDrawer={toggleDrawer} />
+      <NavHeader toggleDrawer={toggleDrawer} />
       <Nav open={open} group={group} toggleDrawer={toggleDrawer} />
       <Box component="main" sx={{ width: '100%', flexGrow: 1, p: { xs: 2, sm: 3 } }}>
-        <Toolbar />
+        <Toolbar sx={{ minHeight: { lg: 0 } }} />
         <Outlet />
       </Box>
     </Box>
